+++ examples/android-github/src/androidTest/kotlin/at/florianschuster/control/androidgithub/search/GithubViewTest.kt
@@ -14,7 +14,7 @@ import androidx.test.espresso.matcher.ViewMatchers.isRoot
 import androidx.test.espresso.matcher.ViewMatchers.withId
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import at.florianschuster.control.androidgithub.R
-import at.florianschuster.control.stub
+import at.florianschuster.control.toStub
 import org.hamcrest.Matcher
 import org.hamcrest.Matchers.not
 import org.junit.Before
@@ -32,8 +32,7 @@ internal class GithubViewTest {
         GithubView.GithubViewModelFactory = object : ViewModelProvider.Factory {
             @Suppress("UNCHECKED_CAST")
             override fun <T : ViewModel> create(modelClass: Class<T>): T {
-                viewModel = GithubViewModel().apply { controller.stub() }
-                return viewModel as T
+                return GithubViewModel().apply { controller.toStub() } as T
             }
         }
         launchFragmentInContainer<GithubView>(themeResId = R.style.Theme_MaterialComponents)
@@ -51,20 +50,20 @@ internal class GithubViewTest {
         // then
         assertEquals(
             GithubViewModel.Action.UpdateQuery(testQuery),
-            viewModel.controller.stub().dispatchedActions.last()
+            viewModel.controller.toStub().dispatchedActions.last()
         )
     }
 
     @Test
     fun whenStateOffersLoadingNextPageThenProgressBarIsShown() {
         // when
-        viewModel.controller.stub().emitState(GithubViewModel.State(loadingNextPage = true))
+        viewModel.controller.toStub().emitState(GithubViewModel.State(loadingNextPage = true))
 
         // then
         onView(withId(R.id.loadingProgressBar)).check(matches(isDisplayed()))
 
         // when
-        viewModel.controller.stub().emitState(GithubViewModel.State(loadingNextPage = false))
+        viewModel.controller.toStub().emitState(GithubViewModel.State(loadingNextPage = false))
 
         // then
         onView(withId(R.id.loadingProgressBar)).check(matches(not(isDisplayed())))