+++ examples/android-github/src/main/kotlin/at/florianschuster/control/androidgithub/search/GithubView.kt
@@ -3,7 +3,6 @@ package at.florianschuster.control.androidgithub.search
 import android.content.Intent
 import android.os.Bundle
 import android.view.View
-import android.widget.EditText
 import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.viewModels